+++ src/TestSuite/DriverTestCase.php
@@ -23,9 +23,8 @@
  */
 namespace DatabaseBackup\TestSuite;
 
-use Cake\Core\Configure;
 use Cake\ORM\TableRegistry;
-use Cake\TestSuite\TestCase;
+use DatabaseBackup\TestSuite\TestCase;
 
 /**
  * DriverTestCase class.
@@ -43,6 +42,11 @@ abstract class DriverTestCase extends TestCase
      * @var \Cake\ORM\Table
      */
     protected $Comments;
+
+    /**
+     * @var object
+     */
+    protected $Driver;
 
     /**
      * @var bool
@@ -79,15 +83,7 @@ abstract class DriverTestCase extends TestCase
     {
         parent::tearDown();
 
-        //Deletes all backups
-        foreach (glob(Configure::read(DATABASE_BACKUP . '.target') . DS . '*') as $file) {
-            //@codingStandardsIgnoreLine
-            @unlink($file);
-        }
-
-        Configure::write(DATABASE_BACKUP . '.connection', 'test');
-
-        unset($this->Articles, $this->Comments);
+        unset($this->Articles, $this->Comments, $this->Driver);
     }
 
     /**
@@ -103,110 +99,10 @@ abstract class DriverTestCase extends TestCase
     }
 
     /**
-     * Internal method to test `export()` and `import()` methods.
-     *
-     * It tests that the backup is properly exported and then imported.
-     * @param object $driverInstance A driver instance
-     * @param string $backup Backup relative path
+     * Test for `_exportExecutable()` method
      * @return void
      */
-    final protected function _testExportAndImport($driverInstance, $backup = 'example.sql')
-    {
-        $backup = $this->getAbsolutePath($backup);
-
-        //Initial records. 3 articles and 6 comments
-        $initial = $this->allRecords();
-        $this->assertEquals(3, count($initial['Articles']));
-        $this->assertEquals(6, count($initial['Comments']));
-
-        //Exports backup
-        $this->assertTrue($driverInstance->export($backup));
-
-        //Deletes article with ID 2 and comment with ID 4
-        $this->Articles->delete($this->Articles->get(2), ['atomic' => false]);
-        $this->Comments->delete($this->Comments->get(4), ['atomic' => false]);
-
-        //Records after delete. 2 articles and 5 comments
-        $afterDelete = $this->allRecords();
-        $this->assertEquals(count($afterDelete['Articles']), count($initial['Articles']) - 1);
-        $this->assertEquals(count($afterDelete['Comments']), count($initial['Comments']) - 1);
-
-        //Imports backup
-        $this->assertTrue($driverInstance->import($backup));
-
-        //Now initial records are the same of final records
-        $final = $this->allRecords();
-        $this->assertEquals($initial, $final);
-
-        //Gets the difference (`$diff`) between records after delete
-        //  (`$deleted`)and records after import (`$final`)
-        $diff = $final;
-
-        foreach ($final as $model => $finalValues) {
-            foreach ($finalValues as $finalKey => $finalValue) {
-                foreach ($afterDelete[$model] as $deletedValue) {
-                    if ($finalValue == $deletedValue) {
-                        unset($diff[$model][$finalKey]);
-                    }
-                }
-            }
-        }
-
-        $this->assertEquals(1, count($diff['Articles']));
-        $this->assertEquals(1, count($diff['Comments']));
-
-        //Difference is article with ID 2 and comment with ID 4
-        $this->assertEquals(2, collection($diff['Articles'])->extract('id')->first());
-        $this->assertEquals(4, collection($diff['Comments'])->extract('id')->first());
-    }
-
-    /**
-     * Test for `getCompression()` method
-     * @return void
-     */
-    abstract public function testGetCompression();
-
-    /**
-     * Test for `getDefaultExtension()` method
-     * @return void
-     */
-    abstract public function testGetDefaultExtension();
-
-    /**
-     * Test for `getExportExecutable()` method
-     * @return void
-     */
-    abstract public function testGetExportExecutable();
-
-    /**
-     * Test for `getExtension()` method
-     * @return void
-     */
-    abstract public function testGetExtension();
-
-    /**
-     * Test for `getImportExecutable()` method
-     * @return void
-     */
-    abstract public function testGetImportExecutable();
-
-    /**
-     * Test for `getValidExtensions()` method
-     * @return void
-     */
-    abstract public function testGetValidExtensions();
-
-    /**
-     * Test for `getValidCompressions()` method
-     * @return void
-     */
-    abstract public function testGetValidCompressions();
-
-    /**
-     * Test for `export()` method
-     * @return void
-     */
-    abstract public function testExport();
+    abstract public function testExportExecutable();
 
     /**
      * Test for `export()` method on failure
@@ -215,10 +111,10 @@ abstract class DriverTestCase extends TestCase
     abstract public function testExportOnFailure();
 
     /**
-     * Test for `import()` method
+     * Test for `_importExecutable()` method
      * @return void
      */
-    abstract public function testImport();
+    abstract public function testImportExecutable();
 
     /**
      * Test for `import()` method on failure
@@ -227,10 +123,175 @@ abstract class DriverTestCase extends TestCase
     abstract public function testImportOnFailure();
 
     /**
+     * Test for `_exportExecutableWithCompression()` method
+     * @return void
+     */
+    public function testExportExecutableWithCompression()
+    {
+        $basicExecutable = $this->invokeMethod($this->Driver, '_exportExecutable');
+
+        //No compression
+        $result = $this->invokeMethod($this->Driver, '_exportExecutableWithCompression', ['backup.sql']);
+        $expected = $basicExecutable . ' > backup.sql 2>/dev/null';
+        $this->assertEquals($expected, $result);
+
+        //Gzip and Bzip2 compressions
+        foreach (['gzip' => 'backup.sql.gz', 'bzip2' => 'backup.sql.bz2'] as $compression => $filename) {
+            $result = $this->invokeMethod($this->Driver, '_exportExecutableWithCompression', [$filename]);
+            $expected = $basicExecutable . ' | ' . $this->getBinary($compression) . ' > ' . $filename . ' 2>/dev/null';
+            $this->assertEquals($expected, $result);
+        }
+    }
+
+    /**
+     * Test for `_importExecutableWithCompression()` method
+     * @return void
+     */
+    public function testImportExecutableWithCompression()
+    {
+        $basicExecutable = $this->invokeMethod($this->Driver, '_importExecutable');
+
+        //No compression
+        $result = $this->invokeMethod($this->Driver, '_importExecutableWithCompression', ['backup.sql']);
+        $expected = $basicExecutable . ' < backup.sql 2>/dev/null';
+        $this->assertEquals($expected, $result);
+
+        //Gzip and Bzip2 compressions
+        foreach (['gzip' => 'backup.sql.gz', 'bzip2' => 'backup.sql.bz2'] as $compression => $filename) {
+            $result = $this->invokeMethod($this->Driver, '_importExecutableWithCompression', [$filename]);
+            $expected = $this->getBinary($compression) . ' -dc ' . $filename . ' | ' . $basicExecutable . ' 2>/dev/null';
+            $this->assertEquals($expected, $result);
+        }
+    }
+
+    /**
+     * Test for `export()` method
+     * @return void
+     * @test
+     */
+    public function testExport()
+    {
+        $backup = $this->getAbsolutePath('example.sql');
+
+        $this->assertTrue($this->Driver->export($backup));
+        $this->assertFileExists($backup);
+    }
+
+    /**
+     * Test for `export()` method. Export is stopped because the
+     *  `beforeExport()` method returns `false`
+     * @return void
+     * @test
+     */
+    public function testExportStoppedByBeforeExport()
+    {
+        $this->Driver = $this->getMockBuilder(get_class($this->Driver))
+            ->setMethods(['beforeExport'])
+            ->setConstructorArgs([$this->getConnection()])
+            ->getMock();
+
+        $this->Driver->method('beforeExport')
+             ->will($this->returnValue(false));
+
+        $backup = $this->getAbsolutePath('example.sql');
+
+        $this->assertFalse($this->Driver->export($backup));
+        $this->assertFileNotExists($backup);
+    }
+
+    /**
+     * Test for `import()` method
+     * @return void
+     * @test
+     */
+    public function testImport()
+    {
+        $backup = $this->getAbsolutePath('example.sql');
+
+        $this->assertTrue($this->Driver->export($backup));
+        $this->assertTrue($this->Driver->import($backup));
+    }
+
+    /**
+     * Test for `import()` method. Import is stopped because the
+     *  `beforeImport()` method returns `false`
+     * @return void
+     * @test
+     */
+    public function testImportStoppedByBeforeExport()
+    {
+        $this->Driver = $this->getMockBuilder(get_class($this->Driver))
+            ->setMethods(['beforeImport'])
+            ->setConstructorArgs([$this->getConnection()])
+            ->getMock();
+
+        $this->Driver->method('beforeImport')
+             ->will($this->returnValue(false));
+
+        $backup = $this->getAbsolutePath('example.sql');
+
+        $this->assertTrue($this->Driver->export($backup));
+        $this->assertFalse($this->Driver->import($backup));
+    }
+
+    /**
      * Test for `export()` and `import()` methods.
      *
      * It tests that the backup is properly exported and then imported.
      * @return void
+     * @test
      */
-    abstract public function testExportAndImport();
+    public function testExportAndImport()
+    {
+        foreach (VALID_EXTENSIONS as $extension) {
+            $this->loadAllFixtures();
+
+            $backup = $this->getAbsolutePath(sprintf('example.%s', $extension));
+
+            //Initial records. 3 articles and 6 comments
+            $initial = $this->allRecords();
+            $this->assertEquals(3, count($initial['Articles']));
+            $this->assertEquals(6, count($initial['Comments']));
+
+            //Exports backup
+            $this->assertTrue($this->Driver->export($backup));
+
+            //Deletes article with ID 2 and comment with ID 4
+            $this->Articles->delete($this->Articles->get(2), ['atomic' => false]);
+            $this->Comments->delete($this->Comments->get(4), ['atomic' => false]);
+
+            //Records after delete. 2 articles and 5 comments
+            $afterDelete = $this->allRecords();
+            $this->assertEquals(count($afterDelete['Articles']), count($initial['Articles']) - 1);
+            $this->assertEquals(count($afterDelete['Comments']), count($initial['Comments']) - 1);
+
+            //Imports backup
+            $this->assertTrue($this->Driver->import($backup));
+
+            //Now initial records are the same of final records
+            $final = $this->allRecords();
+            $this->assertEquals($initial, $final);
+
+            //Gets the difference (`$diff`) between records after delete
+            //  (`$deleted`)and records after import (`$final`)
+            $diff = $final;
+
+            foreach ($final as $model => $finalValues) {
+                foreach ($finalValues as $finalKey => $finalValue) {
+                    foreach ($afterDelete[$model] as $deletedValue) {
+                        if ($finalValue == $deletedValue) {
+                            unset($diff[$model][$finalKey]);
+                        }
+                    }
+                }
+            }
+
+            $this->assertEquals(1, count($diff['Articles']));
+            $this->assertEquals(1, count($diff['Comments']));
+
+            //Difference is article with ID 2 and comment with ID 4
+            $this->assertEquals(2, collection($diff['Articles'])->extract('id')->first());
+            $this->assertEquals(4, collection($diff['Comments'])->extract('id')->first());
+        }
+    }
 }